+++ docs/changelog.rst
@@ -18,6 +18,9 @@ New Features
 - ``plot_free_energy`` now accepts a ``return_data`` flag that will return
   the data used for the free energy plot(#78).
 
+- Added a new function ``plot_trace2d`` that plots the time evolution of a 2D numpy array
+  using a colorbar to map the time (#108).
+
 Improvements
 ~~~~~~~~~~~~